pyproject.toml
[project]
name = "petopeta"
version = "0.1.1"
description = ""
readme = "README.md"
packages = [{ include = "backend" }]
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "soupsieve>=2.7",
    "langchain>=0.3.0,<0.4.0",
    "langsmith[langsmith-pyo3]==0.3.3rc0",
    "langchain-google-genai>=2.1.2,<3.0.0",
    "langchain-core>=0.3.10,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-openai>=0.3.12,<0.4.0",
    "langchain-anthropic>=0.3.10,<0.4.0",
    "langchain-weaviate>=0.0.3,<0.1.0",
    "langchain-postgres>=0.0.14",
    "langgraph>=0.3.21,<0.4.0",
    "beautifulsoup4>=4.12.2",
    "weaviate-client>=4.0.0",
    "lxml>=4.9.3",
    "voyageai>=0.1.4",
    "pillow>=10.2.0",
    "psycopg>=3.2.6",
    "langchain-xai>=0.2.3",
    "tavily-python>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "certifi>=2025.1.31",
    "langgraph-sdk>=0.1.61,<0.2.0",
    "pytest>=7.3.0",
    "ruff>=0.2.2",
    "pandas>=2.2.2",
    "langgraph-cli>=0.2.7",
    "langgraph-cli[inmem]>=0.2.7",
]
